requirements.txt
absl-py==0.11.0
asn1crypto==0.24.0
astunparse==1.6.3
backcall==0.2.0
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
contextvars==2.4
cryptography==2.1.4
dask==2021.3.0
decorator==5.0.9
distributed==2021.3.0
dominate==2.6.0
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
HeapDict==1.0.1
idna==2.6
immutables==0.15
importlib-metadata==3.1.1
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jsonpatch==1.32
jsonpointer==2.1
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.3.3
msgpack==1.0.2
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.18
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
Pygments==2.9.0
pyxdg==0.25
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.1.0
SecretStorage==2.3.1
six==1.15.0
sortedcontainers==2.3.0
tblib==1.7.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.1
termcolor==1.1.0
toml==0.10.2
toolz==0.11.1
torch==1.4.0
torchfile==0.1.0
torchvision==0.5.0
tornado==6.1
tqdm==4.61.1
traitlets==4.3.3
urllib3==1.26.2
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==0.58.0
Werkzeug==1.0.1
wget==3.2
wrapt==1.12.1
zict==2.0.0
zipp==3.4.0
